fix: 수익률 출력 방식 수정

+++ src/main/java/lotto/view/LottoOutputView.java
@@ -9,7 +9,7 @@ public class LottoOutputView {
     private static final String OUTPUT_LOTTO_RESULT_MESSAGE = "당첨 통계\n---";
     private static final String OUTPUT_LOTTO_RANK_RESULT_MESSAGE = "%d개 일치 (%s원) - %d개\n";
     private static final String OUTPUT_LOTTO_SECOND_RANK_RESULT_MESSAGE = "%d개 일치, 보너스 볼 일치 (%s원) - %d개\n";
-    private static final String OUTPUT_LOTTO_RATE_OF_RETURN_MESSAGE = "총 수익률은 %f%%입니다.";
+    private static final String OUTPUT_LOTTO_RATE_OF_RETURN_MESSAGE = "총 수익률은 %.1f%%입니다.";
 
     public void printPurchaseAmountMessage(int purchaseAmountInput) {
         System.out.printf(OUTPUT_PURCHASE_AMOUNT_MESSAGE, purchaseAmountInput);
@@ -33,7 +33,7 @@ public class LottoOutputView {
         System.out.printf(OUTPUT_LOTTO_RANK_RESULT_MESSAGE, lottoRank.getCount(), lottoRank.getPrizeInfo(), result);
     }
 
-    public void printLottoRateOfReturnMessage(float rate) {
+    public void printLottoRateOfReturnMessage(double rate) {
         System.out.printf(OUTPUT_LOTTO_RATE_OF_RETURN_MESSAGE, rate);
     }
 }
